+++ src/utils/const/formElements.js
@@ -2,7 +2,7 @@ export const TEXTAREA_DEFAULT_ROW = 4;
 
 export const ATTRIBUTE_TITLE = "title";
 
-export const TYPE_SUBMIT = "submit";
-export const TYPE_BUTTON = "button";
+export const BUTTON_TYPE_SUBMIT = "submit";
+export const BUTTON_TYPE_BUTTON = "button";
 
 export const TASK_DATE_FORMAT = "YYYY-MM-DD";
