--- src/utils/const/index.js
@@ -1,47 +1,0 @@
-import {
-  ALT_TAG_LOGO,
-  ALT_TAG_ICON_SEARCH,
-  ALT_TAG_ICON_EDIT,
-  ALT_TAG_ICON_DELETE,
-  ALT_TAG_ICON_CHECK,
-  ICON_LOGO,
-  ICON_SEARCH,
-  ICON_DELETE,
-  ICON_EDIT,
-  ICON_CHECK,
-} from "utils/const/images";
-import {
-  TEXTAREA_DEFAULT_ROW,
-  ATTRIBUTE_TITLE,
-  TYPE_SUBMIT,
-  TYPE_BUTTON,
-  TASK_DATE_FORMAT,
-} from "utils/const/formElements";
-import { ERROR_MESSAGE_TASK_TITLE } from "utils/const/errorMessages";
-import {
-  TASK_FILTER_ALL,
-  TASK_FILTER_INCOMPLETED,
-  TASK_FILTER_COMPLETED,
-} from "./tasks";
-
-export {
-  ALT_TAG_LOGO,
-  ALT_TAG_ICON_SEARCH,
-  ALT_TAG_ICON_EDIT,
-  ALT_TAG_ICON_DELETE,
-  ALT_TAG_ICON_CHECK,
-  ICON_LOGO,
-  ICON_SEARCH,
-  ICON_DELETE,
-  ICON_EDIT,
-  ICON_CHECK,
-  TEXTAREA_DEFAULT_ROW,
-  ATTRIBUTE_TITLE,
-  TYPE_SUBMIT,
-  TYPE_BUTTON,
-  TASK_DATE_FORMAT,
-  ERROR_MESSAGE_TASK_TITLE,
-  TASK_FILTER_ALL,
-  TASK_FILTER_INCOMPLETED,
-  TASK_FILTER_COMPLETED,
-};
